+++ packages/hono-takibi/src/generator/zod/helper/property-schema.ts
@@ -1,6 +1,7 @@
 import { zodToOpenAPI } from '../../../helper/zod-to-openapi.js'
 import type { Schema } from '../../../openapi/index.js'
-import { isArrayWithSchemaReference } from '../../../validator/index.js'
+import { isArrayWithSchemaReference } from '../../../utils/index.js'
+
 import { zod } from '../index.js'
 import { arrayReferenceSchema } from './array-reference-schema.js'
 import { referenceSchema } from './reference-schema.js'
@@ -34,10 +35,8 @@ export function propertySchema(schema: Schema): string {
   if (Boolean(schema.$ref) === true) {
     return referenceSchema(schema)
   }
-
   if (isArrayWithSchemaReference(schema)) {
     return arrayReferenceSchema(schema)
   }
-
   return zodToOpenAPI(zod(schema), schema)
 }
